move TestConfig into separate file
